networks moved out of contrib dir mirror, into classifiers
