+++ sktime_dl/classifiers/deeplearning_based/dl4tsc/mlp.py
@@ -1,5 +1,5 @@
-# Fully convolutional neural network, adapted from the implementation from Fawaz et. al
-# https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/fcn.py
+# Multi layer perceptron, adapted from the implementation from Fawaz et. al
+# https://github.com/hfawaz/dl-4-tsc/blob/master/classifiers/mlp.py
 #
 # Network originally proposed by:
 #
@@ -19,10 +19,10 @@ import numpy as np
 import pandas as pd
 
 from sktime.utils.validation import check_X_y
-from sktime_dl.contrib.deeplearning_based.basenetwork import BaseDeepLearner
+from sktime_dl.classifiers.deeplearning_based.basenetwork import BaseDeepLearner
 
 
-class FCN(BaseDeepLearner):
+class MLP(BaseDeepLearner):
 
     def __init__(self, dim_to_use=0, rand_seed=0, verbose=False):
         self.verbose = verbose
@@ -36,7 +36,7 @@ class FCN(BaseDeepLearner):
         self.history = None
 
         # predefined
-        self.nb_epochs = 2000
+        self.nb_epochs = 5000
         self.batch_size = 16
         self.callbacks = None
 
@@ -46,29 +46,27 @@ class FCN(BaseDeepLearner):
     def build_model(self, input_shape, nb_classes, **kwargs):
         input_layer = keras.layers.Input(input_shape)
 
-        conv1 = keras.layers.Conv1D(filters=128, kernel_size=8, padding='same')(input_layer)
-        conv1 = keras.layers.normalization.BatchNormalization()(conv1)
-        conv1 = keras.layers.Activation(activation='relu')(conv1)
+        # flatten/reshape because when multivariate all should be on the same axis
+        input_layer_flattened = keras.layers.Flatten()(input_layer)
 
-        conv2 = keras.layers.Conv1D(filters=256, kernel_size=5, padding='same')(conv1)
-        conv2 = keras.layers.normalization.BatchNormalization()(conv2)
-        conv2 = keras.layers.Activation('relu')(conv2)
+        layer_1 = keras.layers.Dropout(0.1)(input_layer_flattened)
+        layer_1 = keras.layers.Dense(500, activation='relu')(layer_1)
 
-        conv3 = keras.layers.Conv1D(128, kernel_size=3, padding='same')(conv2)
-        conv3 = keras.layers.normalization.BatchNormalization()(conv3)
-        conv3 = keras.layers.Activation('relu')(conv3)
+        layer_2 = keras.layers.Dropout(0.2)(layer_1)
+        layer_2 = keras.layers.Dense(500, activation='relu')(layer_2)
 
-        gap_layer = keras.layers.pooling.GlobalAveragePooling1D()(conv3)
+        layer_3 = keras.layers.Dropout(0.2)(layer_2)
+        layer_3 = keras.layers.Dense(500, activation='relu')(layer_3)
 
-        output_layer = keras.layers.Dense(nb_classes, activation='softmax')(gap_layer)
+        output_layer = keras.layers.Dropout(0.3)(layer_3)
+        output_layer = keras.layers.Dense(nb_classes, activation='softmax')(output_layer)
 
         model = keras.models.Model(inputs=input_layer, outputs=output_layer)
 
-        model.compile(loss='categorical_crossentropy', optimizer=keras.optimizers.Adam(),
+        model.compile(loss='categorical_crossentropy', optimizer=keras.optimizers.Adadelta(),
                       metrics=['accuracy'])
 
-        reduce_lr = keras.callbacks.ReduceLROnPlateau(monitor='loss', factor=0.5, patience=50,
-                                                      min_lr=0.0001)
+        reduce_lr = keras.callbacks.ReduceLROnPlateau(monitor='loss', factor=0.5, patience=200, min_lr=0.1)
 
         # file_path = self.output_directory + 'best_model.hdf5'
         # model_checkpoint = keras.callbacks.ModelCheckpoint(filepath=file_path, monitor='loss',
@@ -105,3 +103,4 @@ class FCN(BaseDeepLearner):
 
         self.history = self.model.fit(X, y_onehot, batch_size=self.batch_size, epochs=self.nb_epochs,
                                       verbose=self.verbose, callbacks=self.callbacks)
+
