+++ sktime_dl/classifiers/deeplearning_based/dl4tsc/twiesn.py
@@ -24,7 +24,7 @@ from sklearn.linear_model import Ridge
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import accuracy_score
 
-from sktime_dl.contrib.deeplearning_based.basenetwork import BaseDeepLearner
+from sktime_dl.classifiers.deeplearning_based.basenetwork import BaseDeepLearner
 
 
 # class Classifier_TWIESN:
